chore: move ReportPortal formatting to create-report
