+++ cmd/root.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/redhat-appstudio/qe-tools/cmd/reportportal"
 	"os"
 
 	"github.com/spf13/cobra"
@@ -46,7 +45,6 @@ func init() {
 	rootCmd.AddCommand(prowjob.ProwjobCmd)
 	rootCmd.AddCommand(coffeebreak.CoffeeBreakCmd)
 	rootCmd.AddCommand(sendslackmessage.SendSlackMessageCmd)
-	rootCmd.AddCommand(reportportal.PrepareRPCmd)
 }
 
 // initConfig reads in config file and ENV variables if set.
